refactor(ratelimit): extract algorithms into a transport-agnostic module
The token bucket, sliding-window counter, and bucket store move to
aquilia/_ratelimit.py, a leaf module with no framework imports. They were
private to the HTTP rate-limit middleware, but the WebSocket limiter needs
the same enforcement math and reaching into another subsystem's underscore
names is a dependency nothing polices. middleware_ext/rate_limit.py keeps
the historical _TokenBucket/_SlidingWindowCounter/_BucketStore aliases so
existing imports resolve unchanged.

Also guards TokenBucket.consume against a zero refill rate. A limit=0 rule
gives refill_rate=0.0, and the deficit branch divided by it. Both the HTTP
and socket limiters route through here, so the guard belongs in the shared
function rather than at each call site; such a bucket now reports a finite
retry-after instead of raising ZeroDivisionError.

BucketStore gains discard() so callers that know a key is dead -- a socket
disconnect -- can release it immediately instead of waiting out the TTL.
